Added an employee flow, mostly for the Android project.

+++ SqlServerAPI/Classes/ProductOrders.cs
@@ -2,7 +2,7 @@
 {
     public class ProductOrders
     {
-        public Int16 orderQuantity { get; set; }
+        public int orderQuantity { get; set; }
         public string? orderName { get; set; }
         public decimal listPrice { get; set; }
